+++ ChangeLog.md
@@ -1,5 +1,7 @@
 # ChangeLog
 
+- Feature: Improve stability for general and fluorescent markers gridding process.
+- Feature: Support for general marker chip (e.g. ZION (Z1C), Kenai (K1B)) gridding process.
 - Feature: Parameterize the statistics window size ratio for evluating the probe intensity whose coefficient of variation is the smallest.
 - Feature: Change the logger shown type for chip id, fov id and channel information in fov_ag.hpp (trace -> info).